pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hexis"
version = "0.1.0"
description = "Hexis memory system (Postgres + pgvector + AGE) with Python tooling."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "pydantic>=2.4.2",
  "asyncpg>=0.29.0",
  "psycopg2-binary>=2.9.9",
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "numpy>=1.24.0",
  "tiktoken>=0.5.1",
  "openai>=1.3.0",
  "pgvector>=0.2.0",
  "reflex==0.8.24.post1",
  "mcp",
]

[project.scripts]
hexis = "apps.hexis_cli:main"
hexis-worker = "apps.worker:main"
hexis-mcp = "apps.hexis_mcp_server:main"
hexis-channels = "services.channel_worker:main"

[project.optional-dependencies]
anthropic = [
  "anthropic>=0.18.0",
]
channels = [
  "discord.py>=2.3.0",
  "python-telegram-bot>=21.0",
  "slack-bolt>=1.18.0",
  "slack-sdk>=3.27.0",
  "matrix-nio>=0.24.0",
]
dev = [
  "pytest>=7.4.3",
  "pytest-asyncio>=0.21.1",
  "pytest-mock>=3.11.1",
  "pytest-cov>=4.1.0",
  "httpx>=0.25.0",
  "tenacity>=8.2.3",
  "selenium>=4.19.0",
  "black>=23.11.0",
  "isort>=5.12.0",
  "mypy>=1.7.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "core*", "services*", "channels*", "plugins*", "skills*"]

[tool.setuptools.package-data]
services = ["prompts/*.md"]
